+++ group/asg/variables.tf
@@ -29,8 +29,8 @@ variable "subnets" {
   type = list(string)
 }
 
-## LC parameters
-variable "lc_id" {
+## LT parameters
+variable "lt_id" {
   type = string
 }
 
@@ -77,7 +77,7 @@ variable "placement_group" {
 }
 
 variable "protect_from_scale_in" {
-  type = string
+  type    = string
   default = "false"
 }